seperated ProtestLatestPictures, ProtestDates
