+++ src/Containers/SafeRender.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function SafeRender({ validObj }) {
+  return validObj && { children };
+}
+
+export default SafeRender;
